Track a finger position
This new code determines if the index finger dtected is extended or not.
To achieve this, The program looks for the location of the tip and base of the finger, it then estimates the middle point beteen these two points. In this location we make a 20 pixel wide circle. If the middle part of the finger is within the cicle then in theory the finger is extended. how ever if either the tip or the base of the finger are also in the circle the the finger is not extended.